Create READ ME
In this repository you find various examples of Python 3 features and some tricks.
Typically, these are what I have used in a class or lecture.
The code is generally commented. I have not altered it from the way I used in class.
So, occasional references to nowhere could be found in the included explanations which you can safely ignore.